Refactor backup and library item upload processes to use same internal-api streaming proxy as cover upload; Refactor upload code for DRY
