+++ src/app/internal-api/items/[id]/cover/route.ts
@@ -1,5 +1,5 @@
 import { getServerBaseUrl } from '@/lib/api'
-import { attachRefreshedSessionCookies, fetchBackendWithCookieRefresh } from '@/lib/serverBackendProxy'
+import { proxyMultipartUpload } from '@/lib/proxyMultipartUpload'
 import { cookies } from 'next/headers'
 import { NextRequest, NextResponse } from 'next/server'
 
@@ -14,30 +14,9 @@ export async function POST(request: NextRequest, { params }: { params: Promise<{
   const cookieStore = await cookies()
 
   try {
-    const contentType = request.headers.get('content-type')
-    if (!contentType?.includes('multipart/form-data')) {
-      return NextResponse.json({ error: 'Invalid content type' }, { status: 400 })
-    }
-
-    if (!request.body) {
-      return NextResponse.json({ error: 'Empty request body' }, { status: 400 })
-    }
-
     const baseUrl = getServerBaseUrl()
     const backendUrl = `${baseUrl}/api/items/${id}/cover`
-
-    const result = await fetchBackendWithCookieRefresh(backendUrl, cookieStore, {
-      method: 'POST',
-      body: request.body,
-      headers: { 'Content-Type': contentType }
-    })
-
-    if (!result.ok) {
-      return attachRefreshedSessionCookies(NextResponse.json({ error: result.error }, { status: result.status }), result.refreshedTokens)
-    }
-
-    const data = await result.upstream.json()
-    return attachRefreshedSessionCookies(NextResponse.json(data), result.refreshedTokens)
+    return await proxyMultipartUpload(request, backendUrl, cookieStore, { forwardJsonResponse: true })
   } catch (error) {
     console.error('[CoverUploadProxy] Error uploading cover:', error)
     return NextResponse.json({ error: 'Failed to upload cover' }, { status: 500 })
