+++ src/types/upload.ts
@@ -1,0 +1,5 @@
+export interface UploadProgressInfo {
+  percent: number
+  loaded: number
+  total: number
+}
